LPS-51335 Move the logic to the service layer to avoid doing multiple calls to the service (add and update) and make it transactional.
