+++ portal-impl/src/com/liferay/portlet/blogs/action/EditEntryAction.java
@@ -14,13 +14,10 @@
 
 package com.liferay.portlet.blogs.action;
 
-import com.liferay.portal.kernel.editor.util.EditorConstants;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.sanitizer.SanitizerException;
 import com.liferay.portal.kernel.servlet.SessionErrors;
 import com.liferay.portal.kernel.servlet.taglib.ui.ImageSelector;
@@ -28,16 +25,13 @@ import com.liferay.portal.kernel.upload.LiferayFileItemException;
 import com.liferay.portal.kernel.upload.UploadException;
 import com.liferay.portal.kernel.util.CalendarFactoryUtil;
 import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.HttpUtil;
-import com.liferay.portal.kernel.util.MimeTypesUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.model.TrashedModel;
-import com.liferay.portal.portletfilerepository.PortletFileRepositoryUtil;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextFactory;
@@ -62,18 +56,11 @@ import com.liferay.portlet.blogs.model.BlogsEntry;
 import com.liferay.portlet.blogs.service.BlogsEntryLocalServiceUtil;
 import com.liferay.portlet.blogs.service.BlogsEntryServiceUtil;
 import com.liferay.portlet.documentlibrary.FileSizeException;
-import com.liferay.portlet.documentlibrary.service.DLAppServiceUtil;
 import com.liferay.portlet.trash.util.TrashUtil;
-
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
 
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -96,7 +83,6 @@ import org.apache.struts.action.ActionMapping;
  * @author Juan Fernández
  * @author Zsolt Berentey
  * @author Levente Hudák
- * @author Roberto Díaz
  */
 public class EditEntryAction extends PortletAction {
 
@@ -357,16 +343,6 @@ public class EditEntryAction extends PortletAction {
 		}
 	}
 
-	protected String getAttachmentLink(
-			FileEntry fileEntryAttachment, ThemeDisplay themeDisplay)
-		throws PortalException {
-
-		String attachmentURL = PortletFileRepositoryUtil.getPortletFileEntryURL(
-			themeDisplay, fileEntryAttachment, StringPool.BLANK);
-
-		return "<img src=\"" + attachmentURL + "\" />";
-	}
-
 	protected String getSaveAndContinueRedirect(
 			PortletConfig portletConfig, ActionRequest actionRequest,
 			BlogsEntry entry, String redirect)
@@ -403,28 +379,6 @@ public class EditEntryAction extends PortletAction {
 		portletURL.setWindowState(actionRequest.getWindowState());
 
 		return portletURL.toString();
-	}
-
-	protected List<FileEntry> getTempFileEntryAttachments(String content)
-		throws PortalException {
-
-		List<FileEntry> fileEntryAttachments = new ArrayList<>();
-
-		Pattern pattern = Pattern.compile(
-			EditorConstants.DATA_IMAGE_ID_ATTRIBUTE + "=.(\\d+)");
-
-		Matcher matcher = pattern.matcher(content);
-
-		while (matcher.find()) {
-			String fileEntryId = matcher.group(1);
-
-			FileEntry fileEntryAttachment = DLAppServiceUtil.getFileEntry(
-				Long.valueOf(fileEntryId));
-
-			fileEntryAttachments.add(fileEntryAttachment);
-		}
-
-		return fileEntryAttachments;
 	}
 
 	protected void subscribe(ActionRequest actionRequest) throws Exception {
@@ -499,38 +453,8 @@ public class EditEntryAction extends PortletAction {
 		}
 	}
 
-	protected String updateContentAttachmentLinks(
-			BlogsEntry entry, String content,
-			List<FileEntry> tempFileEntryAttachments, ThemeDisplay themeDisplay)
-		throws PortalException {
-
-		List<FileEntry> fileEntryAttachments = new ArrayList<>(
-			tempFileEntryAttachments.size());
-
-		for (FileEntry tempAttachment : tempFileEntryAttachments) {
-			FileEntry fileEntryAttachment =
-				BlogsEntryServiceUtil.addEntryAttachment(
-					entry.getGroupId(), entry.getEntryId(),
-					tempAttachment.getTitle(),
-					tempAttachment.getMimeType(),
-					tempAttachment.getContentStream());
-
-			fileEntryAttachments.add(fileEntryAttachment);
-
-			content = StringUtil.replace(
-				content,
-				getAttachmentLink(tempAttachment, themeDisplay),
-				getAttachmentLink(fileEntryAttachment, themeDisplay));
-		}
-
-		return content;
-	}
-
 	protected Object[] updateEntry(ActionRequest actionRequest)
 		throws Exception {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
 
 		long entryId = ParamUtil.getLong(actionRequest, "entryId");
 
@@ -601,9 +525,6 @@ public class EditEntryAction extends PortletAction {
 		ServiceContext serviceContext = ServiceContextFactory.getInstance(
 			BlogsEntry.class.getName(), actionRequest);
 
-		List<FileEntry> tempFileEntryAttachments =
-			getTempFileEntryAttachments(content);
-
 		if (entryId <= 0) {
 
 			// Add entry
@@ -618,29 +539,10 @@ public class EditEntryAction extends PortletAction {
 			AssetPublisherUtil.addAndStoreSelection(
 				actionRequest, BlogsEntry.class.getName(), entry.getEntryId(),
 				-1);
-
-			if (entry != null && !tempFileEntryAttachments.isEmpty()) {
-				content = updateContentAttachmentLinks(
-					entry, content, tempFileEntryAttachments, themeDisplay);
-
-				entry = BlogsEntryServiceUtil.updateEntry(
-					entry.getEntryId(), title, subtitle, description, content,
-					displayDateMonth, displayDateDay, displayDateYear,
-					displayDateHour, displayDateMinute, allowPingbacks,
-					allowTrackbacks, trackbacks, coverImageImageSelector,
-					smallImageImageSelector, serviceContext);
-			}
 		}
 		else {
 
 			// Update entry
-
-			entry = BlogsEntryLocalServiceUtil.getEntry(entryId);
-
-			if (!tempFileEntryAttachments.isEmpty()) {
-				content = updateContentAttachmentLinks(
-					entry, content, tempFileEntryAttachments, themeDisplay);
-			}
 
 			boolean sendEmailEntryUpdated = ParamUtil.getBoolean(
 				actionRequest, "sendEmailEntryUpdated");
@@ -653,6 +555,8 @@ public class EditEntryAction extends PortletAction {
 
 			serviceContext.setAttribute(
 				"emailEntryUpdatedComment", emailEntryUpdatedComment);
+
+			entry = BlogsEntryLocalServiceUtil.getEntry(entryId);
 
 			String tempOldUrlTitle = entry.getUrlTitle();
