+++ portal-impl/src/com/liferay/portlet/blogs/service/impl/BlogsEntryLocalServiceImpl.java
@@ -17,6 +17,7 @@ package com.liferay.portlet.blogs.service.impl;
 import com.liferay.portal.kernel.bean.BeanReference;
 import com.liferay.portal.kernel.comment.CommentManager;
 import com.liferay.portal.kernel.dao.orm.QueryDefinition;
+import com.liferay.portal.kernel.editor.util.EditorConstants;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
@@ -93,12 +94,15 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.Serializable;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletURL;
@@ -238,7 +242,18 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 		entry.setUrlTitle(
 			getUniqueUrlTitle(entryId, title, null, serviceContext));
 		entry.setDescription(description);
+
+		List<FileEntry> tempFileEntryAttachments = getTempFileEntryAttachments(
+			content);
+
+		if (!tempFileEntryAttachments.isEmpty()) {
+			content = updateContentAttachmentLinks(
+				groupId, userId, entry.getEntryId(), content,
+				tempFileEntryAttachments);
+		}
+
 		entry.setContent(content);
+
 		entry.setDisplayDate(displayDate);
 		entry.setAllowPingbacks(allowPingbacks);
 		entry.setAllowTrackbacks(allowTrackbacks);
@@ -290,22 +305,6 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 		}
 
 		return startWorkflowInstance(userId, entry, serviceContext);
-	}
-
-	@Override
-	public FileEntry addEntryAttachment(
-			long groupId, long userId, long entryId, String fileName,
-			String mimeType, InputStream is)
-		throws PortalException {
-
-		BlogsEntry entry = getEntry(entryId);
-
-		Folder folder = entry.addAttachmentsFolder();
-
-		return PortletFileRepositoryUtil.addPortletFileEntry(
-			groupId, userId, BlogsEntry.class.getName(), entry.getEntryId(),
-			PortletKeys.BLOGS, folder.getFolderId(), is, fileName, mimeType,
-			true);
 	}
 
 	@Override
@@ -1228,7 +1227,18 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 		entry.setUrlTitle(
 			getUniqueUrlTitle(entryId, title, oldUrlTitle, serviceContext));
 		entry.setDescription(description);
+
+		List<FileEntry> tempFileEntryAttachments = getTempFileEntryAttachments(
+			content);
+
+		if (!tempFileEntryAttachments.isEmpty()) {
+			content = updateContentAttachmentLinks(
+				entry.getGroupId(), userId, entry.getEntryId(), content,
+				tempFileEntryAttachments);
+		}
+
 		entry.setContent(content);
+
 		entry.setDisplayDate(displayDate);
 		entry.setAllowPingbacks(allowPingbacks);
 		entry.setAllowTrackbacks(allowTrackbacks);
@@ -1542,6 +1552,21 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 		}
 	}
 
+	protected FileEntry addEntryAttachment(
+			long groupId, long userId, long entryId, String fileName,
+			String mimeType, InputStream is)
+		throws PortalException {
+
+		BlogsEntry entry = getEntry(entryId);
+
+		Folder folder = entry.addAttachmentsFolder();
+
+		return PortletFileRepositoryUtil.addPortletFileEntry(
+			groupId, userId, BlogsEntry.class.getName(), entry.getEntryId(),
+			PortletKeys.BLOGS, folder.getFolderId(), is, fileName, mimeType,
+			true);
+	}
+
 	protected long addSmallImageFileEntry(
 			long userId, long groupId, long entryId, String mimeType,
 			String title, InputStream is)
@@ -1568,6 +1593,15 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 	protected void deleteDiscussion(BlogsEntry entry) throws PortalException {
 		commentManager.deleteDiscussion(
 			BlogsEntry.class.getName(), entry.getEntryId());
+	}
+
+	protected String getAttachmentLink(FileEntry fileEntryAttachment)
+		throws PortalException {
+
+		String attachmentURL = PortletFileRepositoryUtil.getPortletFileEntryURL(
+			null, fileEntryAttachment, StringPool.BLANK);
+
+		return "<img src=\"" + attachmentURL + "\" />";
 	}
 
 	protected String getEntryURL(
@@ -1599,6 +1633,28 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 		portletURL.setParameter("entryId", String.valueOf(entry.getEntryId()));
 
 		return portletURL.toString();
+	}
+
+	protected List<FileEntry> getTempFileEntryAttachments(String content)
+		throws PortalException {
+
+		List<FileEntry> fileEntryAttachments = new ArrayList<>();
+
+		Pattern pattern = Pattern.compile(
+			EditorConstants.DATA_IMAGE_ID_ATTRIBUTE + "=.(\\d+)");
+
+		Matcher matcher = pattern.matcher(content);
+
+		while (matcher.find()) {
+			long fileEntryId = GetterUtil.getLong(matcher.group(1));
+
+			FileEntry fileEntryAttachment =
+				PortletFileRepositoryUtil.getPortletFileEntry(fileEntryId);
+
+			fileEntryAttachments.add(fileEntryAttachment);
+		}
+
+		return fileEntryAttachments;
 	}
 
 	protected String getUniqueUrlTitle(
@@ -1999,6 +2055,31 @@ public class BlogsEntryLocalServiceImpl extends BlogsEntryLocalServiceBaseImpl {
 			entry.getCompanyId(), entry.getGroupId(), userId,
 			BlogsEntry.class.getName(), entry.getEntryId(), entry,
 			serviceContext, workflowContext);
+	}
+
+	protected String updateContentAttachmentLinks(
+			long groupId, long userId, long entryId, String content,
+			List<FileEntry> tempFileEntryAttachments)
+		throws PortalException {
+
+		List<FileEntry> fileEntryAttachments = new ArrayList<>(
+			tempFileEntryAttachments.size());
+
+		for (FileEntry tempAttachment : tempFileEntryAttachments) {
+			FileEntry fileEntryAttachment =
+				addEntryAttachment(
+					groupId, userId, entryId, tempAttachment.getTitle(),
+					tempAttachment.getMimeType(),
+					tempAttachment.getContentStream());
+
+			fileEntryAttachments.add(fileEntryAttachment);
+
+			content = StringUtil.replace(
+				content, getAttachmentLink(tempAttachment),
+				getAttachmentLink(fileEntryAttachment));
+		}
+
+		return content;
 	}
 
 	protected void validate(
